Refactor time utilities and modularize `TimeStarts` and `TimeData` classes
- Moved `TimeStarts` and `TimeData` classes to separate modules for better structure and reusability.
- Renamed `utils/timeUtils.py` to `utils/timeUtilities/timeUtil.py` to reflect modular reorganization.
- Introduced `timeDataClasses.py` and `startAndEndBlocks.py` for clean separation of time-related logic.
- Updated imports across files to use the new structure.
- Removed duplicated and unused logic in `timeUtils`.
- Expanded docstrings and examples for clearer usage guidance.
